about to incorporate much more affine space storage.

+++ src/listTools/ChoosePlanes.java
@@ -35,7 +35,7 @@ public class ChoosePlanes {
      * @param k
      * @return
      */
-    public static long choose(long n, long k) {
+    public static int choose(int n, int k) {
         if (k == 0) return 1;
         return choose(n - 1, k - 1) * n / k;
     }
